+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.5.0 (Next)
 - Add snapshot restoration functionality.
+- Add snapshot streaming compression functionality
 - Migrate `snapshot_cleanup` and `snapshot_path` into new `snapshot_config` HCL2 block.
 - Deprecate environment variable input parameter support.